+++ lib/dd_taoke_sdk_web.dart
@@ -4,48 +4,21 @@ import 'dart:async';
 // of your plugin as a separate package, instead of inlining it in the same
 // package as the core of your plugin.
 // ignore: avoid_web_libraries_in_flutter
-import 'dart:html' as html show window;
+import 'dart:html' as html;
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+
+import 'flutter_dd_dataoke_sdk.dart';
 
 /// A web implementation of the DdTaokeSdk plugin.
 class DdTaokeSdkWeb {
   static void registerWith(Registrar registrar) {
-    final MethodChannel channel = MethodChannel(
-      'flutter_dd_dataoke_sdk',
-      const StandardMethodCodec(),
-      registrar.messenger,
-    );
-
     final pluginInstance = DdTaokeSdkWeb();
-    channel.setMethodCallHandler(pluginInstance.handleMethodCall);
+    FlutterDdDataokeSdk.instance = pluginInstance;
   }
 
-  /// Handles method calls over the MethodChannel of this plugin.
-  /// Note: Check the "federated" architecture for a new way of doing this:
-  /// https://flutter.dev/go/federated-plugins
-  Future<dynamic> handleMethodCall(MethodCall call) async {
-    switch (call.method) {
-      case 'getPlatformVersion':
-        return getPlatformVersion();
-      case 'isWeChatBrowser':
-        return isWeChatBrowser();
-      default:
-        throw PlatformException(
-          code: 'Unimplemented',
-          details: 'dd_taoke_sdk for web doesn\'t implement \'${call.method}\'',
-        );
-    }
-  }
-
-  /// Returns a [String] containing the version of the platform.
-  Future<String> getPlatformVersion() {
-    final version = html.window.navigator.userAgent;
-    return Future.value(version);
-  }
-
+  /// 判断是否为微信浏览器
   Future<bool> isWeChatBrowser() async {
     if (kIsWeb) {
       final ua = html.window.navigator.userAgent.toLowerCase();